feat: stereo calibration and intrinsics refinement system
- Add auto-calibration of stereo camera intrinsics via bundle adjustment
- Implement online intrinsics refinement during VIO
- Add offline post-processing calibration tools
- Support multi-camera calibration with guidance system
- Add rolling shutter detection and compensation
- Includes comprehensive test suite and validation framework
- Adds calibration examples and configuration templates

+++ tests/benchmark_sliding_window_optimization.rs
@@ -1,16 +1,16 @@
-use rs_vio::estimator::{SlidingWindow, Frame};
+use rs_vio::estimator::{Frame, SlidingWindow};
 use rs_vio::feature_tracker::Feature;
 use std::time::Instant;
 
 #[test]
 fn bench_optimization_build() {
     let mut sw = SlidingWindow::new(8);
-    
+
     // Add 8 frames
     for i in 0..8 {
-        let mut frame = Frame::new((i * 100) as i64, i as i32);
+        let mut frame = Frame::new((i * 100) as i64, i);
         frame.is_keyframe = true;
-        
+
         // Add 2000 features per frame to stress test
         for j in 0..2000 {
             // Feature ID: j, present in all frames
@@ -26,6 +26,13 @@ fn bench_optimization_build() {
     let start = Instant::now();
     let (problem, initials) = sw.build_optimization_problem();
     let duration = start.elapsed();
-    println!("BENCHMARK: SlidingWindow::build_optimization_problem took: {:?}", duration);
-    println!("  Variables: {}, Residuals: {}", initials.len(), problem.num_residual_blocks());
+    println!(
+        "BENCHMARK: SlidingWindow::build_optimization_problem took: {:?}",
+        duration
+    );
+    println!(
+        "  Variables: {}, Residuals: {}",
+        initials.len(),
+        problem.num_residual_blocks()
+    );
 }
